perf: Optimize RSE.LateUpdate by patching AudioSource ctor (#42)
Every frame, RocketSoundEnhancement makes a FindObjectsOfType call in
order to discover any new AudioSources that have been created in the
last frame. This is slow.

This commit does away with that by:
- Patching AudioSource::.ctor to add the newly created AudioSource
  object to a list (if there is currently an instance of
  RocketSoundEnhancement active).
- Processing the new items in the list in LateUpdate instead of checking
  all available items.

There might be some further simplifications that can be done here now
that we know that we're only processing new AudioSources. None of them
would have an effect on performance though, so I have left them in just
to be safe.

In my testing this change reduces the time spent in RSE.LateUpdate from
4.4s to 90ms over a 253s launch to orbit.